cleanUnscheduled uses new Path scheduling
Since unscheduled EDProducers can now depend on scheduled modules the
we can now remove all EDProducers from a Path. In addition, if an EDFilter
is marked as cms.ignore, it can also be removed from a Path.

To help add cms.ignore to EDFilters on a Path, we now have the
function ignoreAllFiltersOnPath.