Make ContainerMask work with LazyGetter of DetSetVector
edmNew::DetSetVector::dataSize can update as algorithms access the
cached values in edmNew::DetSetVector::dataSize. ContainerMask queries
that value to enforce that the mask it is given is compatible. Now
we just require that the mask be no longer than dataSize. If an
index beyond that range is requested, it returns false.