Extend the unrunnable schedule checker for a case being missed
The general circular dependency checker has been missing a case where
a module on a Path depends on data from a module which is only on an
EndPath. Special code for this case has been added.

I believe the special case is missed because the circular dependency checker
only finds 'base cycles' and not all cycles in the graph. The full cycles
always contain the 'back edge' for a 'base cycle'. However, in this particular
case the base cycle in question is determined (correctly) to be a false positive.
It is only one of the full cylces containing that same 'back edge' which
actually has an unrunnable schedule.