Make sure loop completes before callback task runs
When processing the end paths the code was not guaranteeing that
the loop completes before the endPathDone task can start. This
did happen when there were many end paths in the job where some
ended very quickly and the thread with the loop happened to have
been paused by the system.