Force untracked parameters for modules into Registry
If an input file contained the same PSet ID (because of identical
tracked parameters) as a module in a process, the untracked
parameters for the module are not available in the Registry. This
caused problems with the deleteEarly system which uses an untracked
parameter. We now force the PSet for a module to be injected into
the Registry.