Do not reset info for the stack trace helper
In the case where we are using multiple EventProcessors in one job (such as when one uses FWCore.PythonFramework.CmsRun.py), make sure we do not reset the information cached for the stack trace helper. In particular, we do not want to create another thread nor do things with the pipes. This change does allow forking to cause the values to reset.