Avoid doing parallizable work in a serial call
The call to processEventAsync is made in the source's serial queue task. The processEventAsync call does a lot of work, all of which can actually be done as a different task. Now processEventAsync immediately launched a new tbb task which just does the work of the original processEventAsync but this time no longer within the timing of the serial queue task. This doubled the scalability of a simple hyper fast configuration.