Change EventSetupRecords to be transient objects
Internally, the EventSetup now uses a generic EventSetupRecordImpl
class to manage the work needed for a Record. All classes that inherit
from EventSetupRecord are now just temporary 'facade' objects (similar to
the edm::Event).
This reduces the number of template classes that need to be instantiated and
will be needed to allow enforcement of 'consumes' within the EventSetup system.